[backend] engine refactoring and optimization (#9122)
