+++ opencti-platform/opencti-graphql/src/database/engine.js
@@ -18,7 +18,6 @@ import {
   INDEX_DELETED_OBJECTS,
   INDEX_DRAFT_OBJECTS,
   INDEX_INTERNAL_OBJECTS,
-  inferIndexFromConceptType,
   isEmptyField,
   isInferredIndex,
   isNotEmptyField,
@@ -27,18 +26,8 @@ import {
   pascalize,
   READ_DATA_INDICES,
   READ_DATA_INDICES_WITHOUT_INTERNAL_WITHOUT_INFERRED,
-  READ_ENTITIES_INDICES,
-  READ_INDEX_INFERRED_ENTITIES,
-  READ_INDEX_INFERRED_RELATIONSHIPS,
   READ_INDEX_INTERNAL_OBJECTS,
-  READ_INDEX_INTERNAL_RELATIONSHIPS,
-  READ_INDEX_STIX_CORE_RELATIONSHIPS,
-  READ_INDEX_STIX_CYBER_OBSERVABLE_RELATIONSHIPS,
-  READ_INDEX_STIX_CYBER_OBSERVABLES,
   READ_INDEX_STIX_DOMAIN_OBJECTS,
-  READ_INDEX_STIX_META_OBJECTS,
-  READ_INDEX_STIX_META_RELATIONSHIPS,
-  READ_INDEX_STIX_SIGHTING_RELATIONSHIPS,
   READ_PLATFORM_INDICES,
   READ_RELATIONSHIPS_INDICES,
   UPDATE_OPERATION_ADD,
@@ -56,14 +45,11 @@ import {
   RELATION_OBJECT_LABEL,
   RELATION_OBJECT_MARKING,
   RELATION_OBJECT_PARTICIPANT,
-  STIX_REF_RELATIONSHIP_TYPES
 } from '../schema/stixRefRelationship';
 import {
   ABSTRACT_BASIC_RELATIONSHIP,
-  ABSTRACT_STIX_REF_RELATIONSHIP,
   BASE_TYPE_RELATION,
   buildRefRelationKey,
-  buildRefRelationSearchKey,
   ENTITY_TYPE_IDENTITY,
   ID_INFERRED,
   ID_INTERNAL,
@@ -76,51 +62,24 @@ import {
 import { isModifiedObject, isUpdatedAtObject, } from '../schema/fieldDataAdapter';
 import { getParentTypes, keepMostRestrictiveTypes } from '../schema/schemaUtils';
 import {
-  ATTRIBUTE_ABSTRACT,
   ATTRIBUTE_ALIASES,
   ATTRIBUTE_ALIASES_OPENCTI,
-  ATTRIBUTE_DESCRIPTION,
-  ATTRIBUTE_DESCRIPTION_OPENCTI,
-  ATTRIBUTE_EXPLANATION,
-  ATTRIBUTE_NAME,
   ENTITY_TYPE_IDENTITY_INDIVIDUAL,
   ENTITY_TYPE_IDENTITY_SYSTEM,
   ENTITY_TYPE_LOCATION_COUNTRY,
-  isStixDomainObject,
-  STIX_ORGANIZATIONS_RESTRICTED,
-  STIX_ORGANIZATIONS_UNRESTRICTED
 } from '../schema/stixDomainObject';
-import { isBasicObject, isStixCoreObject, isStixObject } from '../schema/stixCoreObject';
-import { isBasicRelationship, isStixRelationship } from '../schema/stixRelationship';
-import { isStixCoreRelationship, RELATION_INDICATES, RELATION_PUBLISHES, RELATION_RELATED_TO, STIX_CORE_RELATIONSHIPS } from '../schema/stixCoreRelationship';
+import { isStixObject } from '../schema/stixCoreObject';
+import { isBasicRelationship } from '../schema/stixRelationship';
+import { RELATION_INDICATES, RELATION_PUBLISHES, RELATION_RELATED_TO } from '../schema/stixCoreRelationship';
 import { generateInternalId, INTERNAL_FROM_FIELD, INTERNAL_TO_FIELD } from '../schema/identifier';
-import {
-  BYPASS,
-  computeUserMemberAccessIds,
-  controlUserRestrictDeleteAgainstElement,
-  executionContext,
-  INTERNAL_USERS,
-  isBypassUser,
-  MEMBER_ACCESS_ALL,
-  SYSTEM_USER,
-  userFilterStoreElements
-} from '../utils/access';
+import { controlUserRestrictDeleteAgainstElement, executionContext, SYSTEM_USER, userFilterStoreElements } from '../utils/access';
 import { isSingleRelationsRef, } from '../schema/stixEmbeddedRelationship';
 import { now, runtimeFieldObservableValueScript } from '../utils/format';
-import { ENTITY_TYPE_KILL_CHAIN_PHASE, ENTITY_TYPE_MARKING_DEFINITION, isStixMetaObject } from '../schema/stixMetaObject';
-import { getEntitiesListFromCache, getEntityFromCache } from './cache';
-import { ENTITY_TYPE_MIGRATION_STATUS, ENTITY_TYPE_SETTINGS, ENTITY_TYPE_STATUS, ENTITY_TYPE_USER, isInternalObject } from '../schema/internalObject';
+import { ENTITY_TYPE_KILL_CHAIN_PHASE, ENTITY_TYPE_MARKING_DEFINITION } from '../schema/stixMetaObject';
+import { getEntitiesListFromCache } from './cache';
+import { ENTITY_TYPE_MIGRATION_STATUS, ENTITY_TYPE_STATUS, ENTITY_TYPE_USER } from '../schema/internalObject';
 import { meterManager, telemetry } from '../config/tracing';
-import {
-  isBooleanAttribute,
-  isDateAttribute,
-  isDateNumericOrBooleanAttribute,
-  isNumericAttribute,
-  isObjectAttribute,
-  isObjectFlatAttribute,
-  schemaAttributesDefinition,
-  validateDataBeforeIndexing
-} from '../schema/schema-attributes';
+import { isBooleanAttribute, isDateAttribute, isNumericAttribute, isObjectAttribute, schemaAttributesDefinition, validateDataBeforeIndexing } from '../schema/schema-attributes';
 import { convertTypeToStixType } from './stix-converter';
 import { extractEntityRepresentativeName, extractRepresentative } from './entity-representative';
 import { ENTITY_TYPE_IDENTITY_ORGANIZATION } from '../modules/organization/organization-types';
@@ -147,26 +106,19 @@ import {
   X_OPENCTI_WORKFLOW_ID
 } from '../utils/filtering/filtering-constants';
 import { FilterMode } from '../generated/graphql';
-import {
-  booleanMapping,
-  dateMapping,
-  iAliasedIds,
-  internalId,
-  longStringFormats,
-  numericMapping,
-  shortMapping,
-  shortStringFormats,
-  standardId,
-  textMapping,
-  xOpenctiStixIds
-} from '../schema/attribute-definition';
-import { schemaTypesDefinition } from '../schema/schema-types';
-import { INTERNAL_RELATIONSHIPS, isInternalRelationship } from '../schema/internalRelationship';
-import { isStixSightingRelationship, STIX_SIGHTING_RELATIONSHIP } from '../schema/stixSightingRelationship';
-import { rule_definitions } from '../rules/rules-definition';
 import { buildElasticSortingForAttributeCriteria } from '../utils/sorting';
 import { ENTITY_TYPE_DELETE_OPERATION } from '../modules/deleteOperation/deleteOperation-types';
 import { buildEntityData } from './data-builder';
+import { internalEngineMappingGenerator } from './engine-mapping-generator';
+import {
+  buildDataRestrictions,
+  buildFieldForQuery,
+  buildIdsBoolQuery,
+  buildSubQueryForFilterGroup,
+  buildTypesBoolQuery,
+  computeQueryIndices,
+  elGenerateFullTextSearchShould
+} from './engine-query-builder';
 import { buildDraftFilter, DRAFT_OPERATION_CREATE, DRAFT_OPERATION_DELETE_LINKED, DRAFT_OPERATION_DELETE, DRAFT_OPERATION_UPDATE, isDraftSupportedEntity } from './draft-utils';
 import { controlUserConfidenceAgainstElement } from '../utils/confidence-level';
 import { getDraftContext } from '../utils/draftContext';
@@ -177,8 +129,6 @@ import { isStixCyberObservable } from '../schema/stixCyberObservable';
 const ELK_ENGINE = 'elk';
 const OPENSEARCH_ENGINE = 'opensearch';
 export const ES_MAX_CONCURRENCY = conf.get('elasticsearch:max_concurrency');
-export const ES_DEFAULT_WILDCARD_PREFIX = booleanConf('elasticsearch:search_wildcard_prefix', false);
-export const ES_DEFAULT_FUZZY = booleanConf('elasticsearch:search_fuzzy', false);
 export const ES_INIT_RETRO_MAPPING_MIGRATION = booleanConf('elasticsearch:internal_init_retro_compatible_mapping_migration', false);
 export const ES_MINIMUM_FIXED_PAGINATION = 20; // When really low pagination is better by default
 export const ES_DEFAULT_PAGINATION = conf.get('elasticsearch:default_pagination_result') || 500;
@@ -403,145 +353,6 @@ export const elUpdateByQueryForMigration = elOperationForMigration(elRawUpdateBy
 export const elDeleteByQueryForMigration = elOperationForMigration(elRawDeleteByQuery);
 export const elReindexByQueryForMigration = elOperationForMigration(elRawReindexByQuery);
 
-export const buildDataRestrictions = async (context, user, opts = {}) => {
-  const must = [];
-  // eslint-disable-next-line camelcase
-  const must_not = [];
-  // If internal users of the system, we cancel rights checking
-  if (INTERNAL_USERS[user.id]) {
-    return { must, must_not };
-  }
-  // check user access
-  must.push(...buildUserMemberAccessFilter(user, { includeAuthorities: opts?.includeAuthorities }));
-  // If user have bypass, no need to check restrictions
-  if (!isBypassUser(user)) {
-    // region Handle marking restrictions
-    if (user.allowed_marking.length === 0) {
-      // If user have no marking, he can only access to data with no markings.
-      must_not.push({ exists: { field: buildRefRelationKey(RELATION_OBJECT_MARKING) } });
-    } else {
-      // Markings should be grouped by types for restriction
-      const userGroupedMarkings = R.groupBy((m) => m.definition_type, user.allowed_marking);
-      const allGroupedMarkings = R.groupBy((m) => m.definition_type, user.all_marking);
-      const markingGroups = Object.keys(allGroupedMarkings);
-      const mustNotHaveOneOf = [];
-      for (let index = 0; index < markingGroups.length; index += 1) {
-        const markingGroup = markingGroups[index];
-        const markingsForGroup = allGroupedMarkings[markingGroup].map((i) => i.internal_id);
-        const userMarkingsForGroup = (userGroupedMarkings[markingGroup] || []).map((i) => i.internal_id);
-        // Get all markings the user has no access for this group
-        const res = markingsForGroup.filter((m) => !userMarkingsForGroup.includes(m));
-        if (res.length > 0) {
-          mustNotHaveOneOf.push(res);
-        }
-      }
-      // If use have marking, he can access to data with no marking && data with according marking
-      const mustNotMarkingTerms = [];
-      for (let i = 0; i < mustNotHaveOneOf.length; i += 1) {
-        const markings = mustNotHaveOneOf[i];
-        const should = markings.map((m) => ({ match: { [buildRefRelationSearchKey(RELATION_OBJECT_MARKING)]: m } }));
-        mustNotMarkingTerms.push({
-          bool: {
-            should,
-            minimum_should_match: 1,
-          },
-        });
-      }
-      const markingBool = {
-        bool: {
-          should: [
-            {
-              bool: {
-                must_not: [{ exists: { field: buildRefRelationSearchKey(RELATION_OBJECT_MARKING) } }],
-              },
-            },
-            {
-              bool: {
-                must_not: mustNotMarkingTerms,
-              },
-            },
-          ],
-          minimum_should_match: 1,
-        },
-      };
-      must.push(markingBool);
-    }
-    // endregion
-    // region Handle organization restrictions
-    // If user have organization management role, he can bypass this restriction.
-    // If platform is for specific organization, only user from this organization can access empty defined
-    const settings = await getEntityFromCache(context, user, ENTITY_TYPE_SETTINGS);
-    // We want to exlucde a set of entities from organization restrictions while forcing restrictions for an other set of entities
-    const excludedEntityMatches = {
-      bool: {
-        must: [
-          {
-            bool: { must_not: [{ terms: { 'entity_type.keyword': STIX_ORGANIZATIONS_RESTRICTED } }] }
-          },
-          {
-            bool: {
-              should: [
-                { terms: { 'parent_types.keyword': STIX_ORGANIZATIONS_UNRESTRICTED } },
-                { terms: { 'entity_type.keyword': STIX_ORGANIZATIONS_UNRESTRICTED } }
-              ],
-              minimum_should_match: 1
-            }
-          }
-        ]
-      }
-    };
-    if (settings.platform_organization) {
-      if (user.inside_platform_organization) {
-        // Data are visible independently of the organizations
-        // Nothing to restrict.
-      } else {
-        // Data with Empty granted_refs are not visible
-        // Data with granted_refs users that participate to at least one
-        const should = [excludedEntityMatches];
-        const shouldOrgs = user.allowed_organizations
-          .map((m) => ({ match: { [buildRefRelationSearchKey(RELATION_GRANTED_TO)]: m.internal_id } }));
-        should.push(...shouldOrgs);
-        // User individual or data created by this individual must be accessible
-        if (user.individual_id) {
-          should.push({ match: { 'internal_id.keyword': user.individual_id } });
-          should.push({ match: { [buildRefRelationSearchKey(RELATION_CREATED_BY)]: user.individual_id } });
-        }
-        // For tasks
-        should.push({ match: { 'initiator_id.keyword': user.internal_id } });
-        // Access to authorized members
-        should.push(...buildUserMemberAccessFilter(user, { includeAuthorities: opts?.includeAuthorities, excludeEmptyAuthorizedMembers: true }));
-        // Finally build the bool should search
-        must.push({ bool: { should, minimum_should_match: 1 } });
-      }
-    }
-    // endregion
-  }
-  return { must, must_not };
-};
-
-const buildUserMemberAccessFilter = (user, opts) => {
-  const { includeAuthorities = false, excludeEmptyAuthorizedMembers = false } = opts;
-  const capabilities = user.capabilities.map((c) => c.name);
-  if (includeAuthorities && capabilities.includes(BYPASS)) {
-    return [];
-  }
-  const userAccessIds = computeUserMemberAccessIds(user);
-  // if access_users exists, it should have the user access ids
-  const emptyAuthorizedMembers = { bool: { must_not: { exists: { field: 'authorized_members' } } } };
-  const authorizedFilters = [
-    { terms: { 'authorized_members.id.keyword': [MEMBER_ACCESS_ALL, ...userAccessIds] } },
-  ];
-  if (!excludeEmptyAuthorizedMembers) {
-    authorizedFilters.push(emptyAuthorizedMembers);
-  }
-  if (includeAuthorities) {
-    const roleIds = user.roles.map((r) => r.id);
-    const owners = [...userAccessIds, ...capabilities, ...roleIds];
-    authorizedFilters.push({ terms: { 'authorized_authorities.keyword': owners } });
-  }
-  return [{ bool: { should: authorizedFilters } }];
-};
-
 export const elIndexExists = async (indexName) => {
   const existIndex = await engine.indices.exists({ index: indexName });
   return existIndex === true || oebp(existIndex) === true || existIndex.body === true;
@@ -653,94 +464,9 @@ const elCreateCoreSettings = async () => {
   });
 };
 
-// Engine mapping generation on attributes definition
-const attributeMappingGenerator = (entityAttribute) => {
-  if (entityAttribute.type === 'string') {
-    if (shortStringFormats.includes(entityAttribute.format)) {
-      return shortMapping;
-    }
-    if (longStringFormats.includes(entityAttribute.format)) {
-      return textMapping;
-    }
-    throw UnsupportedError('Cant generated string mapping', { format: entityAttribute.format });
-  }
-  if (entityAttribute.type === 'date') {
-    return dateMapping;
-  }
-  if (entityAttribute.type === 'numeric') {
-    return numericMapping(entityAttribute.precision);
-  }
-  if (entityAttribute.type === 'boolean') {
-    return booleanMapping;
-  }
-  if (entityAttribute.type === 'object') {
-    // For flat object
-    if (entityAttribute.format === 'flat') {
-      return { type: engine instanceof ElkClient ? 'flattened' : 'flat_object' };
-    }
-    // For standard object
-    const properties = {};
-    for (let i = 0; i < entityAttribute.mappings.length; i += 1) {
-      const mapping = entityAttribute.mappings[i];
-      properties[mapping.name] = attributeMappingGenerator(mapping);
-    }
-    const config = { dynamic: 'strict', properties };
-    // Add nested option if needed
-    if (entityAttribute.format === 'nested') {
-      config.type = 'nested';
-    }
-    return config;
-  }
-  throw UnsupportedError('Cant generated mapping', { type: entityAttribute.type });
-};
-const ruleMappingGenerator = () => {
-  const schemaProperties = {};
-  for (let attrIndex = 0; attrIndex < rule_definitions.length; attrIndex += 1) {
-    const rule = rule_definitions[attrIndex];
-    schemaProperties[`i_rule_${rule.id}`] = {
-      dynamic: 'strict',
-      properties: {
-        explanation: shortMapping,
-        dependencies: shortMapping,
-        hash: shortMapping,
-        data: { type: engine instanceof ElkClient ? 'flattened' : 'flat_object' },
-      }
-    };
-  }
-  return schemaProperties;
-};
-const denormalizeRelationsMappingGenerator = () => {
-  const databaseRelationshipsName = [
-    STIX_SIGHTING_RELATIONSHIP,
-    ...STIX_CORE_RELATIONSHIPS,
-    ...INTERNAL_RELATIONSHIPS,
-    ...schemaTypesDefinition.get(ABSTRACT_STIX_REF_RELATIONSHIP)
-  ];
-  const schemaProperties = {};
-  for (let attrIndex = 0; attrIndex < databaseRelationshipsName.length; attrIndex += 1) {
-    const relName = databaseRelationshipsName[attrIndex];
-    schemaProperties[`rel_${relName}`] = {
-      dynamic: 'strict',
-      properties: {
-        internal_id: shortMapping,
-        inferred_id: shortMapping,
-      }
-    };
-  }
-  return schemaProperties;
-};
-const attributesMappingGenerator = () => {
-  const entityAttributes = schemaAttributesDefinition.getAllAttributes();
-  const schemaProperties = {};
-  for (let attrIndex = 0; attrIndex < entityAttributes.length; attrIndex += 1) {
-    const entityAttribute = entityAttributes[attrIndex];
-    schemaProperties[entityAttribute.name] = attributeMappingGenerator(entityAttribute);
-  }
-  return schemaProperties;
-};
-
+// Engine mapping generation
 export const engineMappingGenerator = () => {
-  return { ...attributesMappingGenerator(), ...ruleMappingGenerator(), ...denormalizeRelationsMappingGenerator() };
+  return internalEngineMappingGenerator(engine);
 };
 const computeIndexSettings = (rolloverAlias) => {
   if (engine instanceof ElkClient) {
@@ -1471,55 +1197,6 @@ export const elConvertHits = async (data, opts = {}) => {
   return convertedHits;
 };
 
-export const computeQueryIndices = (indices, typeOrTypes) => {
-  const types = (Array.isArray(typeOrTypes) || isEmptyField(typeOrTypes)) ? typeOrTypes : [typeOrTypes];
-  // If indices are explicitly defined, just rely on the definition
-  if (isEmptyField(indices)) {
-    // If not and have no clue about the expected types, ask for all indices.
-    // Worst case scenario that need to be avoided.
-    if (isEmptyField(types)) {
-      return READ_DATA_INDICES;
-    }
-    // If types are defined we need to infer from them the correct indices
-    return R.uniq(types.map((findType) => {
-      // If defined types are abstract, try to restrict the indices as much as possible
-      if (isAbstract(findType)) {
-        // For objects
-        if (isBasicObject(findType)) {
-          if (isInternalObject(findType)) return [READ_INDEX_INFERRED_ENTITIES, READ_INDEX_INTERNAL_OBJECTS];
-          if (isStixMetaObject(findType)) return [READ_INDEX_INFERRED_ENTITIES, READ_INDEX_STIX_META_OBJECTS];
-          if (isStixDomainObject(findType)) return [READ_INDEX_INFERRED_ENTITIES, READ_INDEX_STIX_DOMAIN_OBJECTS];
-          if (isStixCoreObject(findType)) return [READ_INDEX_INFERRED_ENTITIES, READ_INDEX_STIX_DOMAIN_OBJECTS, READ_INDEX_STIX_CYBER_OBSERVABLES];
-          if (isStixObject(findType)) return [READ_INDEX_INFERRED_ENTITIES, READ_INDEX_STIX_META_OBJECTS, READ_INDEX_STIX_DOMAIN_OBJECTS, READ_INDEX_STIX_CYBER_OBSERVABLES];
-          return READ_ENTITIES_INDICES;
-        }
-        // For relationships
-        if (isBasicRelationship(findType) || STIX_REF_RELATIONSHIP_TYPES.includes(findType)) {
-          if (isInternalRelationship(findType)) return [READ_INDEX_INFERRED_RELATIONSHIPS, READ_INDEX_INTERNAL_RELATIONSHIPS];
-          if (isStixSightingRelationship(findType)) return [READ_INDEX_INFERRED_RELATIONSHIPS, READ_INDEX_STIX_SIGHTING_RELATIONSHIPS];
-          if (isStixCoreRelationship(findType)) return [READ_INDEX_INFERRED_RELATIONSHIPS, READ_INDEX_STIX_CORE_RELATIONSHIPS];
-          if (isStixRefRelationship(findType) || STIX_REF_RELATIONSHIP_TYPES.includes(findType)) {
-            return [READ_INDEX_INFERRED_RELATIONSHIPS, READ_INDEX_STIX_META_RELATIONSHIPS, READ_INDEX_STIX_CYBER_OBSERVABLE_RELATIONSHIPS];
-          }
-          if (isStixRelationship(findType)) {
-            return [READ_INDEX_INFERRED_RELATIONSHIPS, READ_INDEX_STIX_CORE_RELATIONSHIPS, READ_INDEX_STIX_SIGHTING_RELATIONSHIPS, READ_INDEX_STIX_META_RELATIONSHIPS,
-              READ_INDEX_STIX_CYBER_OBSERVABLE_RELATIONSHIPS];
-          }
-          return READ_RELATIONSHIPS_INDICES;
-        }
-        // Fallback
-        throw UnsupportedError('Fail to compute indices for unknown type', { type: findType });
-      }
-      // If concrete type, infer the index from the type
-      if (isBasicObject(findType)) {
-        return [READ_INDEX_INFERRED_ENTITIES, `${inferIndexFromConceptType(findType)}*`];
-      }
-      return [READ_INDEX_INFERRED_RELATIONSHIPS, `${inferIndexFromConceptType(findType)}*`];
-    }).flat());
-  }
-  return indices;
-};
-
 // elFindByIds is not defined to use ordering or sorting (ordering is forced by creation date)
 // It's a way to load a bunch of ids and use in list or map
 export const elFindByIds = async (context, user, ids, opts = {}) => {
@@ -1539,36 +1216,16 @@ export const elFindByIds = async (context, user, ids, opts = {}) => {
   for (let index = 0; index < groupIds.length; index += 1) {
     const mustTerms = [];
     const workingIds = groupIds[index];
-    const idsTermsPerType = [];
-    const elementTypes = [internalId.name, standardId.name, xOpenctiStixIds.name, iAliasedIds.name];
-    for (let indexType = 0; indexType < elementTypes.length; indexType += 1) {
-      const elementType = elementTypes[indexType];
-      const terms = { [`${elementType}.keyword`]: workingIds };
-      idsTermsPerType.push({ terms });
-    }
-    const should = {
-      bool: {
-        should: idsTermsPerType,
-        minimum_should_match: 1,
-      },
-    };
+    const should = buildIdsBoolQuery(workingIds);
     mustTerms.push(should);
     if (types && types.length > 0) {
-      const shouldType = {
-        bool: {
-          should: [
-            { terms: { 'entity_type.keyword': types } },
-            { terms: { 'parent_types.keyword': types } }
-          ],
-          minimum_should_match: 1,
-        },
-      };
+      const shouldType = buildTypesBoolQuery(types);
       mustTerms.push(shouldType);
     }
     const restrictionOptions = { includeAuthorities: true }; // By default include authorized through capabilities
     // If an admin ask for a specific element, there is no need to ask him to explicitly extends his visibility to doing it.
-    const markingRestrictions = await buildDataRestrictions(context, user, restrictionOptions);
-    mustTerms.push(...markingRestrictions.must);
+    const accessRestrictions = await buildDataRestrictions(context, user, restrictionOptions);
+    mustTerms.push(...accessRestrictions.must);
     // Handle draft
     const draftMust = buildDraftFilter(context, user, opts);
     const body = {
@@ -1576,7 +1233,7 @@ export const elFindByIds = async (context, user, ids, opts = {}) => {
       query: {
         bool: {
           must: [...mustTerms, ...draftMust],
-          must_not: markingRestrictions.must_not,
+          must_not: accessRestrictions.must_not,
         },
       },
     };
@@ -1635,622 +1292,9 @@ export const elBatchIds = async (context, user, elements) => {
 };
 
 // region elastic common loader.
-export const specialElasticCharsEscape = (query) => {
-  return query.replace(/([/+|\-*()^~={}[\]:?!"\\])/g, '\\$1');
-};
-
-// Global search attributes are limited
-// Its due to opensearch / elastic limitations
-const BASE_SEARCH_CONNECTIONS = [
-  // Pounds for connections search
-  `connections.${ATTRIBUTE_NAME}^4`,
-  // Add all other attributes
-  'connections.*',
-];
-const BASE_SEARCH_ATTRIBUTES = [
-  // Pounds for attributes search
-  `${ATTRIBUTE_NAME}^5`,
-  `${ATTRIBUTE_ABSTRACT}^5`,
-  `${ATTRIBUTE_EXPLANATION}^5`,
-  `${ID_INTERNAL}^5`,
-  `${ID_STANDARD}^5`,
-  `${IDS_STIX}^5`,
-  `${ATTRIBUTE_DESCRIPTION}^2`,
-  `${ATTRIBUTE_DESCRIPTION_OPENCTI}^2`,
-  // For activities
-  'event_type',
-  'event_scope',
-  'context_data.message',
-  // Add all other attributes
-  'aliases',
-  'x_opencti_aliases',
-  'persona_name',
-  'roles',
-  'objective',
-  'content',
-  'content_mapping',
-  'explanation',
-  'opinion',
-  'x_mitre_id',
-  'x_opencti_threat_hunting',
-  'x_opencti_log_sources',
-  'postal_code',
-  'street_address',
-  'source',
-  'context',
-  'pattern',
-  'path',
-  'value',
-  'display_name',
-  'body',
-  'hashes.MD5',
-  'hashes.SHA-1',
-  'hashes.SHA-256',
-  'hashes.SHA-512',
-  'hashes.SHA3-256',
-  'hashes.SHA3-512',
-  'hashes.SSDEEP',
-  'hashes.SDHASH',
-  'hashes.TLSH',
-  'hashes.LZJD',
-  'url',
-  'subject',
-  'payload_bin',
-  'x_opencti_additional_names',
-  'serial_number',
-  'issuer',
-  'cwd',
-  'command_line',
-  'cpe',
-  'swid',
-  'iban',
-  'bic',
-  'account_number',
-  'card_number',
-  'holder_name',
-  'title',
-  'result_name',
-  'phase_name',
-  'kill_chain_name',
-  'definition',
-  'definition_type',
-  'user_email',
-  'main_entity_name', // deletedOperation
-];
-
-function processSearch(search, args) {
-  const { useWildcardPrefix = ES_DEFAULT_WILDCARD_PREFIX } = args;
-  let decodedSearch;
-  try {
-    decodedSearch = decodeURIComponent(search)
-      .trim();
-  } catch (e) {
-    decodedSearch = search.trim();
-  }
-  let remainingSearch = decodedSearch;
-  const exactSearch = (decodedSearch.match(/"[^"]+"/g) || []) //
-    .filter((e) => isNotEmptyField(e.replace(/"/g, '')
-      .trim()));
-  for (let index = 0; index < exactSearch.length; index += 1) {
-    remainingSearch = remainingSearch.replace(exactSearch[index], '');
-  }
-  const querySearch = [];
-
-  const partialSearch = remainingSearch.replace(/"/g, '')
-    .trim()
-    .split(' ');
-
-  for (let searchIndex = 0; searchIndex < partialSearch.length; searchIndex += 1) {
-    const partialElement = partialSearch[searchIndex];
-    const cleanElement = specialElasticCharsEscape(partialElement);
-    if (isNotEmptyField(cleanElement)) {
-      querySearch.push(`${useWildcardPrefix ? '*' : ''}${cleanElement}*`);
-      if (ES_DEFAULT_FUZZY) {
-        querySearch.push(`${cleanElement}~`);
-      }
-    }
-  }
-  return {
-    exactSearch,
-    querySearch
-  };
-}
-
-export const elGenerateFullTextSearchShould = (search, args = {}) => {
-  const { exactSearch, querySearch } = processSearch(search, args);
-  // Return the elastic search engine expected bool should terms
-  // Build the search for all exact match (between double quotes)
-  const shouldSearch = [];
-  const cleanExactSearch = R.uniq(exactSearch.map((e) => e.replace(/"|http?:/g, '')));
-  shouldSearch.push(
-    ...cleanExactSearch.map((ex) => [
-      {
-        multi_match: {
-          type: 'phrase',
-          query: ex,
-          lenient: true,
-          fields: BASE_SEARCH_ATTRIBUTES,
-        },
-      },
-      {
-        nested: {
-          path: 'connections',
-          query: {
-            bool: {
-              must: [
-                {
-                  multi_match: {
-                    type: 'phrase',
-                    query: ex,
-                    lenient: true,
-                    fields: BASE_SEARCH_CONNECTIONS,
-                  },
-                },
-              ],
-            },
-          },
-        },
-      },
-    ]).flat()
-  );
-  // Build the search for all other fields
-  const searchPhrase = R.uniq(querySearch).join(' ');
-  if (searchPhrase) {
-    shouldSearch.push(...[
-      {
-        query_string: {
-          query: searchPhrase,
-          analyze_wildcard: true,
-          fields: BASE_SEARCH_ATTRIBUTES,
-        },
-      },
-      {
-        multi_match: {
-          type: 'phrase',
-          query: searchPhrase,
-          lenient: true,
-          fields: BASE_SEARCH_ATTRIBUTES,
-        },
-      },
-      {
-        nested: {
-          path: 'connections',
-          query: {
-            bool: {
-              must: [
-                {
-                  query_string: {
-                    query: searchPhrase,
-                    analyze_wildcard: true,
-                    fields: BASE_SEARCH_CONNECTIONS,
-                  },
-                },
-              ],
-            },
-          },
-        },
-      },
-    ]);
-  }
-  return shouldSearch;
-};
-
-export const elGenerateFieldTextSearchShould = (search, arrayKeys, args = {}) => {
-  const { exactSearch, querySearch } = processSearch(search, args);
-  const cleanExactSearch = R.uniq(exactSearch.map((e) => e.replace(/"|http?:/g, '')));
-  const shouldSearch = [];
-  shouldSearch.push(
-    ...cleanExactSearch.map((ex) => [
-      {
-        multi_match: {
-          type: 'phrase',
-          query: ex,
-          lenient: true,
-          fields: arrayKeys,
-        },
-      }
-    ]).flat()
-  );
-  // Build the search for all other fields
-  const searchPhrase = R.uniq(querySearch).join(' ');
-  if (searchPhrase) {
-    shouldSearch.push(...[
-      {
-        query_string: {
-          query: searchPhrase,
-          analyze_wildcard: true,
-          fields: arrayKeys,
-        },
-      },
-      {
-        multi_match: {
-          type: 'phrase',
-          query: searchPhrase,
-          lenient: true,
-          fields: arrayKeys,
-        },
-      }
-    ]);
-  }
-
-  return shouldSearch;
-};
 
 const BASE_FIELDS = ['_index', 'internal_id', 'standard_id', 'sort', 'base_type', 'entity_type',
   'connections', 'first_seen', 'last_seen', 'start_time', 'stop_time'];
-
-const RANGE_OPERATORS = ['gt', 'gte', 'lt', 'lte'];
-
-const buildFieldForQuery = (field) => {
-  return isDateNumericOrBooleanAttribute(field) || field === '_id' || isObjectFlatAttribute(field)
-    ? field
-    : `${field}.keyword`;
-};
-const buildLocalMustFilter = async (validFilter) => {
-  const valuesFiltering = [];
-  const noValuesFiltering = [];
-  const { key, values, nested, operator = 'eq', mode: localFilterMode = 'or' } = validFilter;
-  if (isEmptyField(key)) {
-    throw FunctionalError('A filter key must be defined', { key });
-  }
-  const arrayKeys = Array.isArray(key) ? key : [key];
-  const headKey = R.head(arrayKeys);
-  const dontHandleMultipleKeys = nested || operator === 'nil' || operator === 'not_nil';
-  if (dontHandleMultipleKeys && arrayKeys.length > 1) {
-    throw UnsupportedError('Filter must have only one field', { keys: arrayKeys, operator });
-  }
-  // 01. Handle nested filters
-  // TODO IF KEY is PART OF Rule we need to add extra fields search
-  // TODO Add connections like filters to have native fromId, toId filters handling.
-  // See opencti-front\src\private\components\events\StixSightingRelationships.tsx
-  if (nested) {
-    const nestedMust = [];
-    const nestedMustNot = [];
-    for (let nestIndex = 0; nestIndex < nested.length; nestIndex += 1) {
-      const nestedElement = nested[nestIndex];
-      const parentKey = arrayKeys.at(0);
-      const { key: nestedKey, values: nestedValues, operator: nestedOperator = 'eq' } = nestedElement;
-      const nestedShould = [];
-      const nestedFieldKey = `${parentKey}.${nestedKey}`;
-      if (nestedKey === ID_INTERNAL) {
-        if (nestedOperator === 'nil') {
-          nestedMustNot.push({
-            exists: {
-              field: nestedFieldKey
-            }
-          });
-        } else if (nestedOperator === 'not_nil') {
-          nestedShould.push({
-            exists: {
-              field: nestedFieldKey
-            }
-          });
-        } else if (nestedOperator === 'not_eq') {
-          nestedMustNot.push({ terms: { [`${nestedFieldKey}.keyword`]: nestedValues } });
-        } else { // nestedOperator = 'eq'
-          nestedShould.push({ terms: { [`${nestedFieldKey}.keyword`]: nestedValues } });
-        }
-      } else { // nested key !== internal_id
-        // eslint-disable-next-line no-lonely-if
-        if (nestedOperator === 'nil') {
-          nestedMustNot.push({
-            exists: {
-              field: nestedFieldKey
-            }
-          });
-        } else if (nestedOperator === 'not_nil') {
-          nestedShould.push({
-            exists: {
-              field: nestedFieldKey
-            }
-          });
-        } else {
-          for (let i = 0; i < nestedValues.length; i += 1) {
-            const nestedSearchValue = nestedValues[i].toString();
-            if (nestedOperator === 'wildcard') {
-              nestedShould.push({ query_string: { query: `${nestedSearchValue}`, fields: [nestedFieldKey] } });
-            } else if (nestedOperator === 'not_eq') {
-              nestedMustNot.push({
-                multi_match: {
-                  fields: buildFieldForQuery(nestedFieldKey),
-                  query: nestedSearchValue.toString(),
-                }
-              });
-            } else if (RANGE_OPERATORS.includes(nestedOperator)) {
-              nestedShould.push({ range: { [nestedFieldKey]: { [nestedOperator]: nestedSearchValue } } });
-            } else { // nestedOperator = 'eq'
-              nestedShould.push({
-                multi_match: {
-                  fields: buildFieldForQuery(nestedFieldKey),
-                  query: nestedSearchValue.toString(),
-                }
-              });
-            }
-          }
-        }
-      }
-      const should = {
-        bool: {
-          should: nestedShould,
-          minimum_should_match: localFilterMode === 'or' ? 1 : nestedShould.length,
-        },
-      };
-      nestedMust.push(should);
-    }
-    const nestedQuery = {
-      path: headKey,
-      query: {
-        bool: {
-          must: nestedMust,
-          must_not: nestedMustNot,
-        },
-      },
-    };
-    return { nested: nestedQuery };
-  }
-  // 02. Handle nil and not_nil operators
-  if (operator === 'nil') {
-    const filterDefinition = schemaAttributesDefinition.getAttributeByName(headKey);
-    let valueFiltering = { // classic filters: field doesn't exist
-      bool: {
-        must_not: {
-          exists: {
-            field: headKey
-          }
-        }
-      }
-    };
-    if (filterDefinition?.type === 'string') {
-      if (filterDefinition?.format === 'text') { // text filters: use wildcard
-        valueFiltering = {
-          bool: {
-            must_not: {
-              wildcard: {
-                [headKey]: '*'
-              }
-            },
-          }
-        };
-      } else { // string filters: nil <-> (field doesn't exist) OR (field = empty string)
-        valueFiltering = {
-          bool: {
-            should: [
-              {
-                bool: {
-                  must_not: {
-                    exists: {
-                      field: headKey
-                    }
-                  }
-                }
-              },
-              {
-                term: {
-                  [headKey === '_id' ? headKey : `${headKey}.keyword`]: { value: '' },
-                },
-              },
-            ],
-            minimum_should_match: 1,
-          }
-        };
-      }
-    } else if (filterDefinition?.type === 'date') { // date filters: nil <-> (field doesn't exist) OR (date <= epoch) OR (date >= 5138)
-      valueFiltering = {
-        bool: {
-          should: [
-            {
-              bool: {
-                must_not: {
-                  exists: {
-                    field: headKey
-                  }
-                }
-              }
-            },
-            { range: { [headKey]: { lte: '1970-01-01T01:00:00.000Z' } } },
-            { range: { [headKey]: { gte: '5138-11-16T09:46:40.000Z' } } }
-          ],
-          minimum_should_match: 1,
-        }
-      };
-    }
-    valuesFiltering.push(valueFiltering);
-  } else if (operator === 'not_nil') {
-    const filterDefinition = schemaAttributesDefinition.getAttributeByName(headKey);
-    let valueFiltering = { // classic filters: field exists
-      exists: {
-        field: headKey
-      }
-    };
-    if (filterDefinition?.type === 'string') {
-      if (filterDefinition?.format === 'text') { // text filters: use wildcard
-        valueFiltering = {
-          bool: {
-            must: {
-              wildcard: {
-                [headKey]: '*'
-              }
-            },
-          }
-        };
-      } else { // other filters: not_nil <-> (field exists) AND (field != empty string)
-        valueFiltering = {
-          bool: {
-            must: [
-              {
-                exists: {
-                  field: headKey
-                }
-              },
-              {
-                bool: {
-                  must_not: {
-                    term: {
-                      [headKey === '_id' ? headKey : `${headKey}.keyword`]: { value: '' },
-                    },
-                  },
-                }
-              }
-            ],
-          }
-        };
-      }
-    } else if (filterDefinition?.type === 'date') { // date filters: not_nil <-> (field exists) AND (date > epoch) AND (date < 5138)
-      valueFiltering = {
-        bool: {
-          must: [
-            {
-              exists: {
-                field: headKey
-              }
-            },
-            { range: { [headKey]: { gt: '1970-01-01T01:00:00.000Z' } } },
-            { range: { [headKey]: { lt: '5138-11-16T09:46:40.000Z' } } }
-          ],
-        }
-      };
-    }
-    valuesFiltering.push(valueFiltering);
-  }
-  // 03. Handle values according to the operator
-  if (operator !== 'nil' && operator !== 'not_nil') {
-    for (let i = 0; i < values.length; i += 1) {
-      if (values[i] === 'EXISTS') {
-        if (arrayKeys.length > 1) {
-          throw UnsupportedError('Filter must have only one field', { keys: arrayKeys });
-        }
-        valuesFiltering.push({ exists: { field: headKey } });
-      } else if (operator === 'eq' || operator === 'not_eq') {
-        const targets = operator === 'eq' ? valuesFiltering : noValuesFiltering;
-        targets.push({
-          multi_match: {
-            fields: arrayKeys.map((k) => buildFieldForQuery(k)),
-            query: values[i].toString(),
-          },
-        });
-      } else if (operator === 'match') {
-        valuesFiltering.push({
-          multi_match: {
-            fields: arrayKeys,
-            query: values[i].toString(),
-          },
-        });
-      } else if (operator === 'wildcard') {
-        valuesFiltering.push({
-          query_string: {
-            query: `"${values[i].toString()}"`,
-            fields: arrayKeys,
-          },
-        });
-      } else if (operator === 'contains' || operator === 'not_contains') {
-        const targets = operator === 'contains' ? valuesFiltering : noValuesFiltering;
-        const val = specialElasticCharsEscape(values[i].toString());
-        targets.push({
-          query_string: {
-            query: `*${val.replace(/\s/g, '\\ ')}*`,
-            analyze_wildcard: true,
-            fields: arrayKeys.map((k) => `${k}.keyword`),
-          },
-        });
-      } else if (operator === 'starts_with' || operator === 'not_starts_with') {
-        const targets = operator === 'starts_with' ? valuesFiltering : noValuesFiltering;
-        const val = specialElasticCharsEscape(values[i].toString());
-        targets.push({
-          query_string: {
-            query: `${val.replace(/\s/g, '\\ ')}*`,
-            analyze_wildcard: true,
-            fields: arrayKeys.map((k) => `${k}.keyword`),
-          },
-        });
-      } else if (operator === 'ends_with' || operator === 'not_ends_with') {
-        const targets = operator === 'ends_with' ? valuesFiltering : noValuesFiltering;
-        const val = specialElasticCharsEscape(values[i].toString());
-        targets.push({
-          query_string: {
-            query: `*${val.replace(/\s/g, '\\ ')}`,
-            analyze_wildcard: true,
-            fields: arrayKeys.map((k) => `${k}.keyword`),
-          },
-        });
-      } else if (operator === 'script') {
-        valuesFiltering.push({
-          script: {
-            script: values[i].toString()
-          },
-        });
-      } else if (operator === 'search') {
-        const shouldSearch = elGenerateFieldTextSearchShould(values[i].toString(), arrayKeys);
-        const bool = {
-          bool: {
-            should: shouldSearch,
-            minimum_should_match: 1,
-          },
-        };
-        valuesFiltering.push(bool);
-      } else {
-        if (arrayKeys.length > 1) {
-          throw UnsupportedError('Filter must have only one field', { keys: arrayKeys });
-        }
-        valuesFiltering.push({ range: { [headKey]: { [operator]: values[i] } } }); // range operators
-      }
-    }
-  }
-  // 04. Push the values
-  if (valuesFiltering.length > 0) {
-    return {
-      bool: {
-        should: valuesFiltering,
-        minimum_should_match: localFilterMode === 'or' ? 1 : valuesFiltering.length,
-      },
-    };
-  }
-  if (noValuesFiltering.length > 0) {
-    return {
-      bool: {
-        should: noValuesFiltering.map((o) => ({
-          bool: {
-            must_not: [o]
-          }
-        })),
-        minimum_should_match: localFilterMode === 'or' ? 1 : noValuesFiltering.length,
-      },
-    };
-  }
-  throw UnsupportedError('Invalid filter configuration', validFilter);
-};
-
-const buildSubQueryForFilterGroup = async (context, user, inputFilters) => {
-  const { mode = 'and', filters = [], filterGroups = [] } = inputFilters;
-  const localMustFilters = [];
-  // Handle filterGroups
-  for (let index = 0; index < filterGroups.length; index += 1) {
-    const group = filterGroups[index];
-    if (isFilterGroupNotEmpty(group)) {
-      const subQuery = await buildSubQueryForFilterGroup(context, user, group);
-      if (subQuery) { // can be null
-        localMustFilters.push(subQuery);
-      }
-    }
-  }
-  // Handle filters
-  for (let index = 0; index < filters.length; index += 1) {
-    const filter = filters[index];
-    const isValidFilter = filter?.values || filter?.nested?.length > 0;
-    if (isValidFilter) {
-      const localMustFilter = await buildLocalMustFilter(filter);
-      localMustFilters.push(localMustFilter);
-    }
-  }
-  if (localMustFilters.length > 0) {
-    return {
-      bool: {
-        should: localMustFilters,
-        minimum_should_match: mode === 'or' ? 1 : localMustFilters.length,
-      }
-    };
-  }
-  return null;
-};
 
 // If filter key = entity_type, we should also handle parent_types
 // Example: filter = {mode: 'or', operator: 'eq', key: ['entity_type'], values: ['Report', 'Stix-Cyber-Observable']}
@@ -2849,25 +1893,29 @@ const elQueryBodyBuilder = async (context, user, options) => {
   const searchAfter = after ? cursorToOffset(after) : undefined;
   let ordering = [];
   const { includeAuthorities = false } = options;
-  // Handle marking restrictions
-  const markingRestrictions = await buildDataRestrictions(context, user, { includeAuthorities });
-  const accessMust = markingRestrictions.must;
-  const accessMustNot = markingRestrictions.must_not;
+  // Handle data access restrictions (marking, organziation sharing, authorized members)
+  const accessRestrictions = await buildDataRestrictions(context, user, { includeAuthorities });
+  const accessMust = accessRestrictions.must;
+  const accessMustNot = accessRestrictions.must_not;
   const mustFilters = [];
+  // build ids query
+  if (ids.length > 0) {
+    const shouldIds = buildIdsBoolQuery(ids);
+    mustFilters.push(shouldIds);
+  }
+  // build types query
+  if (types !== null && types.length > 0) {
+    const shouldType = buildTypesBoolQuery(types);
+    mustFilters.push(shouldType);
+  }
   // Add special keys to filters
   const specialFiltersContent = [];
-  if (ids.length > 0 || startDate || endDate || (types !== null && types.length > 0)) {
-    if (ids.length > 0) {
-      specialFiltersContent.push({ key: IDS_FILTER, values: ids });
-    }
+  if (startDate || endDate) {
     if (startDate) {
       specialFiltersContent.push({ key: dateAttribute || 'created_at', values: [startDate], operator: intervalInclude ? 'gte' : 'gt' });
     }
     if (endDate) {
       specialFiltersContent.push({ key: dateAttribute || 'created_at', values: [endDate], operator: intervalInclude ? 'lte' : 'lt' });
-    }
-    if (types !== null && types.length > 0) {
-      specialFiltersContent.push({ key: TYPE_FILTER, values: R.flatten(types) });
     }
   }
   const completeFilters = specialFiltersContent.length > 0 ? {
@@ -3365,7 +2413,7 @@ export const elLoadBy = async (context, user, field, value, type = null, indices
 export const elAttributeValues = async (context, user, field, opts = {}) => {
   const { orderMode = 'asc', search } = opts;
   const first = opts.first ?? ES_DEFAULT_PAGINATION;
-  const markingRestrictions = await buildDataRestrictions(context, user);
+  const accessRestrictions = await buildDataRestrictions(context, user);
   const must = [];
   if (isNotEmptyField(search) && search.length > 0) {
     const shouldSearch = elGenerateFullTextSearchShould(search);
@@ -3377,12 +2425,12 @@ export const elAttributeValues = async (context, user, field, opts = {}) => {
     };
     must.push(bool);
   }
-  must.push(...markingRestrictions.must);
+  must.push(...accessRestrictions.must);
   const body = {
     query: {
       bool: {
         must,
-        must_not: markingRestrictions.must_not,
+        must_not: accessRestrictions.must_not,
       },
     },
     aggs: {
