+++ opencti-platform/opencti-graphql/tests/01-unit/domain/search-test.js
@@ -1,5 +1,5 @@
 import { expect, it } from 'vitest';
-import { elGenerateFullTextSearchShould, specialElasticCharsEscape } from '../../../src/database/engine';
+import { elGenerateFullTextSearchShould, specialElasticCharsEscape } from '../../../src/database/engine-query-builder';
 import { isNotEmptyField } from '../../../src/database/utils';
 
 const parse = (search) => {
